cpukit/posix/src/mmap.c: Correct comparisons of different signedness
The GCC warning -Wsign-compare flagged these instances of comparing
signed and unsigned types. A common pair was size_t and int.

+++ cpukit/posix/src/mmap.c
@@ -286,7 +286,7 @@ void *mmap(
        */
       r = read( fildes, mapping->addr, len );
 
-      if ( r != len ) {
+      if ( r != (ssize_t) len ) {
         mmap_mappings_lock_release( );
         if ( !map_fixed ) {
           free( mapping->addr );
